+++ lib/isolator/adapters/background_jobs/active_job.rb
@@ -4,7 +4,7 @@ Isolator.isolate :active_job,
   target: ActiveJob::Base,
   method_name: :enqueue,
   exception_class: Isolator::BackgroundJobError,
-  details_message: ->(obj, _args) {
+  details_message: ->(obj) {
     "#{obj.class.name}" \
     "#{obj.arguments.any? ? " (#{obj.arguments.join(", ")})" : ""}"
   }