+++ lib/isolator/adapters/mailers/mail.rb
@@ -3,7 +3,7 @@
 Isolator.isolate :mailer, target: Mail::Message,
                           method_name: :deliver,
                           exception_class: Isolator::MailerError,
-                          details_message: ->(obj, _args) {
+                          details_message: ->(obj) {
                             "From: #{obj.from}\n" \
                             "To: #{obj.to}\n" \
                             "Subject: #{obj.subject}"