+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## master
 
+- Support keyword arguments to isolated method in Ruby 3.0. ([@Mange][])
 - Raise an error when an ignore file does not parse to a hash. ([@bobbymcwho][])
 - Log all filtered backtrace lines to the logger ([@bobbymcwho][])
 - Add support for removing dynamic adapters. ([@Mange][])